appdatvexe/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
idna==3.2
jwcrypto==1.0
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
wrapt==1.12.1
